feat(order): refactor order-related services and models for improved clarity and functionality

+++ backend/src/main/java/com/dev/backend/dto/order/OrderResponse.java
@@ -36,7 +36,7 @@ public class OrderResponse {
     private Integer voucherAmount;
     private Integer shippingFee;
 
-    private Long total;
+    private Integer total;
 
     private LocalDate createdAt;
 }
